No Last-Modified on listfiles (it would go backwards on file removal) Prevent nginx from stripping ETags when gzipping

+++ server/src/fcgi/fcgi-actions-block.c
@@ -241,7 +241,7 @@ static int cb_inuse_number(void *ctx, const char *s, size_t l)
     }
 
     if (yactx->state == CB_IU_BLOCKSIZE) {
-        DEBUG("blocksize: %d", nnumb);
+        DEBUG("blocksize: %lld", (long long)nnumb);
         yactx->meta.blocksize = nnumb;
         yactx->replica = 0;
         yactx->state = CB_IU_VALUES;
